move NSDataExt.kt to korlibs-platform
